Algorithm II: Day 6 BFS / DFS
